Move UI and plugin runtime helpers into services
Interface diffs:

- import "github.com/valon-technologies/gestalt/server/internal/ui"

+ import "github.com/valon-technologies/gestalt/server/services/ui"

- import "github.com/valon-technologies/gestalt/server/internal/adminui"

+ import "github.com/valon-technologies/gestalt/server/services/ui/adminui"

- import "github.com/valon-technologies/gestalt/server/internal/pluginruntime"

+ import "github.com/valon-technologies/gestalt/server/services/runtimehost/pluginruntime"

Examples:

handler, err := ui.DirHandler(assetRoot)

admin := adminui.EmbeddedHandler(adminui.Options{BrandHref: "/"})

runtimeProvider := pluginruntime.NewLocalProvider()
