+++ gestaltd/internal/bootstrap/plugin_runtime.go
@@ -7,8 +7,8 @@ import (
 	"sync"
 
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/pluginruntime"
 	"github.com/valon-technologies/gestalt/server/services/runtimehost"
+	"github.com/valon-technologies/gestalt/server/services/runtimehost/pluginruntime"
 )
 
 type pluginRuntimeRegistry struct {
