+++ gestaltd/internal/bootstrap/plugin_runtime_plan.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/pluginruntime"
+	"github.com/valon-technologies/gestalt/server/services/runtimehost/pluginruntime"
 )
 
 type RuntimeHostServiceAccess string
