+++ gestaltd/internal/server/mounted_ui.go
@@ -13,14 +13,14 @@ import (
 	"slices"
 	"strings"
 
-	"github.com/valon-technologies/gestalt/server/internal/adminui"
 	"github.com/valon-technologies/gestalt/server/internal/config"
 	"github.com/valon-technologies/gestalt/server/internal/providerpkg"
-	"github.com/valon-technologies/gestalt/server/internal/ui"
 	"github.com/valon-technologies/gestalt/server/services/identity/principal"
 	"github.com/valon-technologies/gestalt/server/services/invocation"
 	"github.com/valon-technologies/gestalt/server/services/observability/metricutil"
 	"github.com/valon-technologies/gestalt/server/services/providerdev"
+	"github.com/valon-technologies/gestalt/server/services/ui"
+	"github.com/valon-technologies/gestalt/server/services/ui/adminui"
 )
 
 const browserLoginPath = "/api/v1/auth/login"
