+++ gestaltd/services/runtimehost/pluginruntime/session_log_writer.go
