+++ gestaltd/services/ui/static_handler.go
